move decoration functionality over
