+++ libs/vscode/nx-config-decoration/project.json
@@ -12,5 +12,5 @@
       }
     }
   },
-  "tags": []
+  "tags": ["type:vscode"]
 }
