+++ libs/vscode/nx-config-decoration/src/lib/get-project-locations.ts
