--- libs/vscode/nx-config-decoration/src/lib/vscode-nx-config-decoration.ts
@@ -1,3 +1,0 @@
-export function vscodeNxConfigDecoration(): string {
-  return 'vscode-nx-config-decoration';
-}
