+++ libs/vscode/nx-config-decoration/src/lib/workspace-codelens-provider.ts
@@ -2,27 +2,18 @@ import {
   CodeLens,
   CodeLensProvider,
   Command,
-  ConfigurationChangeEvent,
-  Disposable,
-  ExtensionContext,
-  languages,
   Range,
   TextDocument,
   Uri,
-  workspace,
 } from 'vscode';
 
 import { buildProjectPath } from '@nx-console/shared/utils';
+import { WorkspaceConfigurationStore } from '@nx-console/vscode/configuration';
 import {
-  GlobalConfigurationStore,
-  WorkspaceConfigurationStore,
-} from '@nx-console/vscode/configuration';
-import {
-  getProjectLocations,
-  ProjectLocations,
-} from './find-workspace-json-target';
-import { getNxWorkspace } from './get-nx-workspace';
-import { getProjectByPath } from './get-project-by-path';
+  getNxWorkspace,
+  getProjectByPath,
+} from '@nx-console/vscode/nx-workspace';
+import { ProjectLocations, getProjectLocations } from './get-project-locations';
 
 export class TargetCodeLens extends CodeLens {
   constructor(
@@ -52,21 +43,6 @@ export class TaskGraphCodeLens extends CodeLens {
 }
 
 export class WorkspaceCodeLensProvider implements CodeLensProvider {
-  /**
-   * CodeLensProvider is disposed and re-registered on setting changes
-   */
-  codeLensProvider: Disposable | null;
-
-  /**
-   * The WorkspaceCodeLensProvider adds clickable nx run targets in the workspace config file.
-   * It is enabled by default and can be disabled with the `enableWorkspaceConfigCodeLens` setting.
-   * @param context instance of ExtensionContext from activate
-   */
-  constructor(private readonly context: ExtensionContext) {
-    this.registerWorkspaceCodeLensProvider(context);
-    this.watchWorkspaceCodeLensConfigChange(context);
-  }
-
   /**
    * Provides a CodeLens set for a matched document
    * @param document a document matched by the pattern passed to registerCodeLensProvider
@@ -209,49 +185,5 @@ export class WorkspaceCodeLensProvider implements CodeLensProvider {
     }
 
     return null;
-  }
-
-  /**
-   * Checks the enableWorkspaceConfigCodeLens setting and registers this as a CodeLensProvider.
-   * @param context instance of ExtensionContext from activate
-   */
-  registerWorkspaceCodeLensProvider(context: ExtensionContext) {
-    const enableWorkspaceConfigCodeLens = GlobalConfigurationStore.instance.get(
-      'enableWorkspaceConfigCodeLens'
-    );
-    if (enableWorkspaceConfigCodeLens) {
-      this.codeLensProvider = languages.registerCodeLensProvider(
-        { pattern: '**/{workspace,project}.json' },
-        this
-      );
-      context.subscriptions.push(this.codeLensProvider);
-    }
-  }
-
-  /**
-   * Watches for settings/configuration changes and enables/disables the CodeLensProvider
-   * @param context instance of ExtensionContext from activate
-   */
-  watchWorkspaceCodeLensConfigChange(context: ExtensionContext) {
-    context.subscriptions.push(
-      workspace.onDidChangeConfiguration((event: ConfigurationChangeEvent) => {
-        // if the `nxConsole` config changes, check enableWorkspaceConfigCodeLens and register or dispose
-        const affectsNxConsoleConfig = event.affectsConfiguration(
-          GlobalConfigurationStore.configurationSection
-        );
-        if (affectsNxConsoleConfig) {
-          const enableWorkspaceConfigCodeLens =
-            GlobalConfigurationStore.instance.get(
-              'enableWorkspaceConfigCodeLens'
-            );
-          if (enableWorkspaceConfigCodeLens && !this.codeLensProvider) {
-            this.registerWorkspaceCodeLensProvider(this.context);
-          } else if (!enableWorkspaceConfigCodeLens && this.codeLensProvider) {
-            this.codeLensProvider.dispose();
-            this.codeLensProvider = null;
-          }
-        }
-      })
-    );
   }
 }
