+++ libs/vscode/nx-workspace/src/index.ts
@@ -1,6 +1,3 @@
-export * from './lib/find-workspace-json-target';
-export * from './lib/reveal-workspace-json';
-export * from './lib/workspace-codelens-provider';
 export * from './lib/stop-daemon';
 export * from './lib/get-nx-workspace';
 export * from './lib/get-project-by-path';
