+++ tsconfig.base.json
@@ -108,7 +108,7 @@
       "@nx-console/vscode/utils": ["libs/vscode/utils/src/index.ts"],
       "@nx-console/vscode/verify": ["libs/vscode/verify/src/index.ts"],
       "@nx-console/vscode/webview": ["libs/vscode/webview/src/index.ts"],
-      "vscode/nx-config-decoration": [
+      "@nx-console/vscode/nx-config-decoration": [
         "libs/vscode/nx-config-decoration/src/index.ts"
       ]
     }
